requirements.txt
python-telegram-bot==20.7
langdetect==1.0.9
matplotlib==3.7.2
pandas==2.0.3
numpy==1.24.3
pandas-ta==0.3.14b0
pycoingecko==3.1.0
requests==2.31.0
python-dotenv==1.0.0